Refactored tests to common class
