+++ tests/danfysik8000.py
@@ -61,19 +61,3 @@ class Danfysik8000Tests(DanfysikCommon, unittest.TestCase):
             self.ca.assert_that_pv_is(ilk_pv, "Interlock")
             self._lewis.backdoor_command(["device", "disable_interlock", ilk_name])
             self.ca.assert_that_pv_is(ilk_pv, "OK")
-
-    @skip_if_recsim("Can not test disconnection in rec sim")
-    def test_GIVEN_device_not_connected_WHEN_voltage_pv_checked_THEN_pv_in_alarm(self):
-        sleep(5)
-        self._lewis.backdoor_set_on_device('comms_initialized', False)
-        self._lewis.backdoor_set_on_device('device_available', False)
-        sleep(10)
-        self.ca.assert_that_pv_alarm_is('VOLT', ChannelAccess.Alarms.INVALID)
-
-    @skip_if_recsim("Can not test disconnection in rec sim")
-    def test_GIVEN_device_not_connected_WHEN_current_pv_checked_THEN_pv_in_alarm(self):
-        sleep(5)
-        self._lewis.backdoor_set_on_device('comms_initialized', False)
-        self._lewis.backdoor_set_on_device('device_available', False)
-        sleep(10)
-        self.ca.assert_that_pv_alarm_is('CURR', ChannelAccess.Alarms.INVALID)
